refactor: move location count info code
A recent PR was merged upstream that refactors the source code. This
commit reconciles this merge conflict.
