+++ src/contexts.ts
@@ -1,7 +1,7 @@
 import React from "react";
-import { DateLocationsPair } from "./types";
+import { DateLocationsPair } from "./booking/BookingDataTypes";
 
 export const DateLocationsPairsContext = React.createContext({
-    setDateLocationsPairs: (_a: DateLocationsPair[]): void => {},
-    dateLocationsPairs: [] as DateLocationsPair[],
+  setDateLocationsPairs: (_a: DateLocationsPair[]): void => {},
+  dateLocationsPairs: [] as DateLocationsPair[],
 });
