+++ src/filterOldDates.tsx
@@ -1,4 +1,4 @@
-import { DateLocationsPair } from "./types";
+import { DateLocationsPair } from "./booking/BookingDataTypes";
 
 const getTime = (string: string): number =>
   parseInt(string.substring(0, 5).replace(/:/g, ""), 10);
