+++ src/utils/math.ts
@@ -1,0 +1,3 @@
+export function sum(array: number[]) {
+  return array.reduce((a, b) => a + b, 0);
+}
