+++ src/walk-in/modal/WalkInModal.tsx
@@ -1,33 +1,23 @@
 import { Button, KIND } from "baseui/button";
 import { Modal } from "baseui/modal";
-import "./App.css";
-import { Instruction, WalkinLocation } from "./getData";
-import { ModalGrid } from "./VaxComponents";
-import { NoticeList, NoticeListItem } from "./NoticeList";
+import { ModalGrid } from "../../VaxComponents";
+import { NoticeList, NoticeListItem } from "../../NoticeList";
 import { FunctionComponent } from "react";
+import { CancelBookingNotice } from "./CancelNotice";
+import { Instruction, WalkInLocation } from "../WalkInData";
+import "../../App.css";
 import { faCar, faWalking } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
 type Props = {
   clearSelectedLocation: () => void;
-  location?: WalkinLocation;
+  location?: WalkInLocation;
 };
 
-const CancelBookingNotice: FunctionComponent<{ className: string }> = ({
-  className,
-}) => (
-  <div className={`cancel-booking-notice ${className}`}>
-    <h4>Please don't double-book</h4>
-    <p>
-      Remember to cancel your original booking if you have one at{" "}
-      <a href="https://bookmyvaccine.nz" target="_blank" rel="noreferrer">
-        bookmyvaccine.nz
-      </a>
-    </p>
-  </div>
-);
-
-const WalkModal = ({ clearSelectedLocation, location }: Props) => {
+const WalkInModal: FunctionComponent<Props> = ({
+  clearSelectedLocation,
+  location,
+}) => {
   const close = () => clearSelectedLocation();
   if (location == null) {
     return null;
@@ -212,4 +202,4 @@ const WalkModal = ({ clearSelectedLocation, location }: Props) => {
   );
 };
 
-export default WalkModal;
+export default WalkInModal;
